host/requirements.txt
crcmod==1.7
hid==1.0.6
pyserial==3.5
setuptools==75.1.0

matplotlib~=3.9.2
numpy~=2.1.3
PyQt5~=5.15.11